Catch exception when acknowledging Google purchase
`Google\Service\Exception` with `reason: concurrentUpdate` is thrown when
second `acknowledge()` call is triggered before Google processed the first
call. This can happen on slow network (api slowed) / or when hermes processes
messages too fast.

If exception is thrown in `DeveloperNotificationReceivedHandler` (hermes
handler for notification received in webhook). Hermes message is processed
again later (we find out that purchase was acknowledged and move on).

But if this exception is thrown in `VerifyPurchaseApiHandler` it breaks
process of purchase verification for user and error is displayed in
application. Acknowledgment is here only for internal use - to let Google
know we processed purchase and we don't need them to retry sending developer
notifications. Skipping acknowledgment for sake of user experience is
better than showing useless error. Worst case scenario is Google sending
developer notification again - we find out it is processed and
acknowledge it.

Note: ReceiptValidator catches all exceptions from Google and envelops them
in own `ReceiptValidator\RunTimeException`. So we need to catch that one
and check previous throwable `\Google\Service\Exception`.

All other exceptions are thrown back.

remp/crm#2385